+++ src/modules/sewerlevelData.js
@@ -2,7 +2,7 @@ import { getTime } from "./measureTime.js";
 import "../utils/envConfig.js";
 import axios from "axios";
 
-export const getApiData = async (req) => {
+const getApiData = async (req) => {
   const measure = getTime();
   const gu = req;
 
@@ -10,10 +10,10 @@ export const getApiData = async (req) => {
   let data = {};
 
   //.env 파일에서 서비스인증키를 불러옵니다.
-  //const key = process.env.SERVICE_KEY;
+  const key = process.env.SERVICE_KEY;
 
   //github 올릴 경우, 빈 값으로 적용
-  const key = "";
+  //const key = "";
 
   let drainLevelUrl = `http://openapi.seoul.go.kr:8088/${key}/json/DrainpipeMonitoringInfo/1/1000/${gu}/${startDate}/${endDate}`;
 
@@ -33,7 +33,7 @@ export const getApiData = async (req) => {
     /**서울시 강우량 정보
      * @param {String} gu_name 구청명(용산)
      */
-    let rainDropUrl = `http://openapi.seoul.go.kr:8088/${key}/json/ListRainfallService/1/100/${encodeURIComponent(
+    let rainDropUrl = `http://openapi.seoul.go.kr:8088/${key}/json/ListRainfallService/1/50/${encodeURIComponent(
       gubnNameData
     )}`;
 
@@ -50,3 +50,5 @@ export const getApiData = async (req) => {
     console.error(err);
   }
 };
+
+export default getApiData;
